syncing from original (#2)
* [temporary] Disable v2ray in CI, mark broken

While we're chasing down what's breaking the v2ray install, default it to off and warn people it's not working. Ideally we would fix the v2ray bug first, but it is blocking all other work. :-(

* Ansible 2.8.4; update to digital_ocean_droplet module; vendor Algo's patched version

First, upgrade to Ansible 2.8.4, and require that version.

The dopy module, which is what's behind the `digital_ocean` Ansible module, has multiple issues with DigitalOcean API changes. It's also deprecated and depends on Python 2.7.

Fortunately, `digital_ocean_droplet` and friends were added to Ansible to replace `digital_ocean`.

Unfortunately, `digital_ocean_droplet` was recently broken by DO. Our friends at Algo ran into this too, and filed a PR. Their proposed module is now in the `library` directory. We'll remove it when Ansible fixes `digital_ocean_droplet`.

*  When EC2 VPC undefined, *omit* ec2:assign_public_ip

If the user has specified a VPC to launch into, we can clearly use `assign_public_ip: true`. But if the VPC is undefined, we shouldn't specify this. And apparently `assign_public_ip: false` is not good enough; we need to *omit* the clause.

Fixes #1509. Fixes #1499.

* ad-blocking via dnsmasq; disable cloudflared in CI

This PR implements DNS-level blocking of ads and trackers. Dnsmasq is configured to return `0.0.0.0` for bad domains.

notracking/hosts-blocklists maintains easy-to-use network block lists. Importantly, it supports both host and domain blacklisting. The lists are refreshed daily.

To reduce the mischief that upstream could cause, files downloaded are rewritten through awk filters. Notably, the domain blocker part ensures there are no directives in the dnsmasq config fragment other than address.

As a reminder: Streisand DNS services are available to authenticated users at `10.10.10.10` (among other addresses).

Cloudflared is now disabled in CI. I'm not thrilled with this, but runs with the `cloudflared` target simply weren't completing. A PR turning it back on is welcome.

* [Temporary] workaround for v2ray/core insecure redirect issue (#1651)

See https://github.com/StreisandEffect/streisand/issues/1642

We clone the v2ray/v2ray-core repository manually to avoid go failing one of the playbook tasks with an "insecure redirect" error. This can be removed when either go releases a new version that ignores unsafe redirects, or Github Pages is able to fix the insecure redirect bug that exists for v2ray.com/core (see https://github.com/StreisandEffect/streisand/issues/1642#issuecomment-529265026)

* Update corban@raunco.co gpg key (#1654)